Autox/requirements.txt
selenium
pytest
pytest-html
pytest-xdist
requests
allure-pytest
webdriver-manager
python-dotenv
pyyaml
loguru
tenacity
pytest-rerunfailures